:pencil: Refresh MainNavigator
