+++ screens/FirstTime/ActivityScreen.jsx
@@ -9,11 +9,11 @@ const ActivityScreen = ({ route }) => {
   const moods = ["Super", "Bien", "Bof", "Mal", "Terrible"];
   const mood = moods[moodIndex];
   const images = [
-    require('../../../assets/img/mood-1.png'),
-    require('../../../assets/img/mood-2.png'),
-    require('../../../assets/img/mood-3.png'),
-    require('../../../assets/img/mood-4.png'),
-    require('../../../assets/img/mood-5.png')
+    require('../../assets/img/mood-1.png'),
+    require('../../assets/img/mood-2.png'),
+    require('../../assets/img/mood-3.png'),
+    require('../../assets/img/mood-4.png'),
+    require('../../assets/img/mood-5.png')
   ];
   const selectedImage = images[moodIndex];
 
@@ -22,7 +22,7 @@ const ActivityScreen = ({ route }) => {
       <View style={styles.topContent}>
         <TouchableOpacity onPress={() => navigation.goBack()}>
             <Image
-                source={require('../../../assets/img/left-arrow.png')}
+                source={require('../../assets/img/left-arrow.png')}
                 style={styles.arrow}
                 resizeMode="contain"
             />
@@ -30,7 +30,7 @@ const ActivityScreen = ({ route }) => {
         <Text style={styles.progressText}>2/4</Text>
         <TouchableOpacity onPress={() => navigation.goBack()}>
             <Image
-                source={require('../../../assets/img/cross.png')}
+                source={require('../../assets/img/cross.png')}
                 style={styles.cross}
                 resizeMode="contain"
             />
