+++ screens/FirstTime/MoodScreen.jsx
@@ -2,14 +2,14 @@
 import React, { useEffect, useState } from 'react';
 import { View, Text, Image, StyleSheet, Alert, FlatList, TouchableOpacity } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
-import Button from '../../../components/Button';
+import Button from '../../components/Button';
 
 const images = [
-  require('../../../assets/img/mood-1.png'),
-  require('../../../assets/img/mood-2.png'),
-  require('../../../assets/img/mood-3.png'),
-  require('../../../assets/img/mood-4.png'),
-  require('../../../assets/img/mood-5.png')
+  require('../../assets/img/mood-1.png'),
+  require('../../assets/img/mood-2.png'),
+  require('../../assets/img/mood-3.png'),
+  require('../../assets/img/mood-4.png'),
+  require('../../assets/img/mood-5.png')
 ];
 
 const texts = [
@@ -71,7 +71,7 @@ const MoodScreen = () => {
         <View style={styles.topContent}>
             <TouchableOpacity onPress={() => navigation.goBack()}>
                 <Image
-                    source={require('../../../assets/img/left-arrow.png')}
+                    source={require('../../assets/img/left-arrow.png')}
                     style={styles.arrow}
                     resizeMode="contain"
                 />
@@ -79,7 +79,7 @@ const MoodScreen = () => {
             <Text style={styles.progressText}>1/4</Text>
             <TouchableOpacity onPress={() => navigation.goBack()}>
                 <Image
-                    source={require('../../../assets/img/cross.png')}
+                    source={require('../../assets/img/cross.png')}
                     style={styles.cross}
                     resizeMode="contain"
                 />
@@ -91,14 +91,14 @@ const MoodScreen = () => {
           <View style={styles.dateContainer}>
             <View style={styles.dateTimeContainer}>
                 <Image
-                source={require('../../../assets/img/calendar.png')}
+                source={require('../../assets/img/calendar.png')}
                 style={styles.icon}
                 resizeMode="contain"
                 />
                 <Text style={styles.dateTimeText}>{currentDate}</Text>
                 <View style={{marginRight: 40}} />
                 <Image
-                source={require('../../../assets/img/clock.png')}
+                source={require('../../assets/img/clock.png')}
                 style={styles.icon}
                 resizeMode="contain"
                 />
